Update n3-rule-based-machine-reasoning.markdown
Untill Consistency check for user defined restrictions.

Sentence before (A complete reasoning example on transitivity for the 'is part of' property for text and text structures as prosodic entities is given in the repository, also involving a series of other RDFS- and OWL-rules. The data and query files are commented.) --> Link?